% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{CoverageAnnotationDecorator}
\alias{CoverageAnnotationDecorator}
\title{Add coverage annotation to statement list}
\usage{
CoverageAnnotationDecorator(stmt.list)
}
\arguments{
\item{stmt.list}{statement list}
}
\value{
annotated statement list
}
\description{
Recursive decorator of statement list that adds annotations needed to measure coverage
}

