% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{DecStatement}
\alias{DecStatement}
\title{Replace function body for coverage measurement}
\usage{
DecStatement(stmt)
}
\arguments{
\item{f}{function object}
}
\description{
Decorate function body to be able to measure coverage
}

