% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{elem_autocov}
\alias{elem_autocov}
\title{Element-wise tail-robust autocovariance matrix estimator}
\usage{
elem_autocov(X, l, est.type = "truncate", tau.mat, block.size, ...)
}
\arguments{
\item{X}{A \code{numeric} p x n matrix of observations.}

\item{l}{A non-negative integer corresponding to the lag of the autocovariance.}

\item{est.type}{A \code{character} string, such as "truncate", "huber" or "MoM", indicating the type of element-wise tail-robust estimator.}

\item{tau.mat}{A \code{numeric} p x p matrix of robustification parameters (larger than 0), when est.type is "truncate" or "huber".}

\item{block.size}{A strictly positive integer scalar corresponding to the block size, when est.type is "MoM".}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} scalar corresponding to the median-of-means estimator.
}
\description{
Compute the element-wise tail-robust matrix estimator.
}
\examples{
set.seed(123)
X = matrix(rt(50*100, 3), nrow = 50, ncol = 100)
gamma1_trunc = elem_autocov(X, l = 1, est.type = 'truncate', tau.mat = matrix(5, nrow = 50, ncol = 50))
gamma1_huber = elem_autocov(X, l = 1, est.type = 'huber', tau.mat = matrix(5, nrow = 50, ncol = 50))
gamma1_MoM = elem_autocov(X, l = 1, est.type = 'MoM', block.size = 10)
dim(gamma1_trunc)
dim(gamma1_huber)
dim(gamma1_MoM)

}
\author{
Haotian Xu
}
