% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{huber_mean}
\alias{huber_mean}
\title{Element-wise adaptive Huber mean estimator}
\usage{
huber_mean(x, tau, ...)
}
\arguments{
\item{x}{A \code{numeric} vector of observations.}

\item{tau}{A \code{numeric} scalar corresponding to the robustification parameter (larger than 0).}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} scalar corresponding to the adaptive Huber mean estimator.
}
\description{
Computes the element-wise adaptive Huber mean estimator
}
\examples{
set.seed(123)
X = rt(100, 3)
mean(X)
huber_mean(X, 6)

}
\author{
Haotian Xu
}
