% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{IsControlFlow}
\alias{IsControlFlow}
\title{Is statment is a control flow statement}
\usage{
IsControlFlow(stmt)
}
\arguments{
\item{stmt}{statement to be checked}
}
\value{
\code{TRUE/FALSE}
}
\description{
Check if statement is a control flow statement that needs special handling for coverage annotations
}

