% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{linear_simu_heter}
\alias{linear_simu_heter}
\title{Simulating Linear process model with heteroscedastic errors}
\usage{
linear_simu_heter(n, mu, M.list, rho, K = 1001, err.dist = "t3", ...)
}
\arguments{
\item{n}{A strictly positive integer corresponding to the length of output series.}

\item{mu}{A \code{numeric} vector of mean of VAR2 model.}

\item{M.list}{A list of p x p matrices, and (rho^(k-1)*M.list[[k]]) be the transition matrix for lag-k term.}

\item{rho}{A scalar in (0,1) indicating the decay rate of the transition matrix for lag-k term as the k grows.}

\item{K}{A strictly positive integer and (K-1) represents the number of lags considered.}

\item{err.dist}{A \code{character} string, such as "normal", "pareto", "lognorm" or "t4", indicating the distribution of innovations.}

\item{...}{Additional arguments.}
}
\value{
A p x n matrix corresponding to the simulated data
}
\description{
Simulate a p dimensional Linear process model with length n. The setting follows APPENDIX D: SIMULATION STUDY in Zhang&Wu(2017)
}
\examples{
n = 100
p = 100
K = 1001
set.seed(123)
M.list = vector(mode = "list", length = K)
for(ll in 1:K){
  M.list[[ll]] = matrix(rnorm(p^2), nrow = p)
}
X = linear_simu(n = n, mu = rep(0, p), M.list = M.list, rho = 0.7, err.dist = "pareto")
dim(X)

}
\author{
Haotian Xu
}
