% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{MoM_mean}
\alias{MoM_mean}
\title{Median-of-means estimator}
\usage{
MoM_mean(x, block.size, ...)
}
\arguments{
\item{x}{A \code{numeric} vector of observations.}

\item{block.size}{A strictly positive integer scalar corresponding to the block size.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} scalar corresponding to the median-of-means estimator.
}
\description{
Compute the classical median-of-means estimator.
}
\examples{
set.seed(123)
X = rt(100, 3)
mean(X)
MoM_mean(X, 10)

}
\author{
Haotian Xu
}
