% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{MonitorCoverageHelper}
\alias{MonitorCoverageHelper}
\title{Create new function with coverage annotation}
\usage{
MonitorCoverageHelper(func.obj, func.name)
}
\arguments{
\item{func.obj}{function to be changed}

\item{func.name}{function name}
}
\description{
Change function body to add statements needed to measure code coverage
}

