% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{optim.tuning}
\alias{optim.tuning}
\title{Selecting the optimal robustification parameter tau for the element-wise truncated or adaptive Huber estimators}
\usage{
optim.tuning(
  X,
  S,
  h = 1,
  length_tau = 5,
  M_est = FALSE,
  start.prob = 0.95,
  ...
)
}
\arguments{
\item{X}{A \code{numeric} pxn matrix containing n observations of a p-dimensional time series.}

\item{S}{A strictly positive integer corresponding to the chosen number of non-overlapping blocks.}

\item{h}{A strictly positive integer corresponding to how many neighboring blocks will be removed.}

\item{length_tau}{A strictly positive integer corresponding to the length of the vector of candidate robustification parameters.}

\item{M_est}{A \code{logical} scalar indicating which truncated estimator will be used (F: truncated mean estimator; T: adaptive Huber M-estimator).}

\item{start.prob}{A \code{numeric} scalar in (0,1) such that the minimum value of the candidate robustification parameters is the empirical "start.prob" quantile of the corresponding coordiante of X.}

\item{...}{Additional arguments.}

\item{max.tau}{A positive \code{numeric} scalar corresponding to the maximum value of the candidate robustification parameters.}
}
\value{
A \code{numeric} vector with each entry corresponds the selected robustification parameter for for each coordinate.
}
\description{
As the title
}
\examples{
n = 100
p = 50
Sigma = diag(1, p)
set.seed(12345)
X = VAR1_simu(n = n, mu = rep(-5, p), skip = 300, Sigma.mat = Sigma, rho = 0.7, err.dist = "pareto")
cv_result = optim.tuning(X, S = 10, h = 1, length_tau = 10, M_est = FALSE)
tau_cv = median(cv_result)
max(abs(apply(X, 1, FUN = function(x){trunc_mean(x, tau_cv)}) - rep(-5,50)))
max(abs(apply(cbind(cv_result, X), 1, FUN = function(x){trunc_mean(x[-1], x[1])}) - rep(-5,50)))
max(abs(apply(X, 1, FUN = function(x){mean(x)}) - rep(-5,50)))
}
\author{
Haotian Xu and Stephane Guerrier
}
