% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{ReportCoverageInfo}
\alias{ReportCoverageInfo}
\title{Report Coverage Information}
\usage{
ReportCoverageInfo()
}
\value{
data.frame with coverage information
}
\description{
Gather coverage information in a data frame and return it. Report contains following information per function -
total number of statements, number of non-executed statement, line coverage percentage.
}

