% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{ReportCoveragePercentage}
\alias{ReportCoveragePercentage}
\title{Report Coverage Percentage Only}
\usage{
ReportCoveragePercentage(cov.env)
}
\value{
numeric with coverage percentage
}
\description{
Report Coverage Percentage Only
}

