% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{ReportEnvironmentCoverage}
\alias{ReportEnvironmentCoverage}
\title{Measure coverage of an environment}
\usage{
ReportEnvironmentCoverage(envir, ..., enclos = parent.frame())
}
\arguments{
\item{envir}{environment to monitor coverage in}

\item{...}{for specific tests separeted by commas}

\item{enclos}{environment where tests should be evaluated}
}
\description{
Monitor coverage of an environment using tests
}

