% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{ReportMissingStatements}
\alias{ReportMissingStatements}
\title{Report Missing Statements}
\usage{
ReportMissingStatements()
}
\value{
list of lists of missing statements by function that coverage was monitored for
}
\description{
Report Missing Statements
}

