% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{ReportPackageCoverage}
\alias{ReportPackageCoverage}
\title{Measure coverage of a package}
\usage{
ReportPackageCoverage(package.dir, ..., test.dir = "tests/")
}
\arguments{
\item{package.dir}{package directory}

\item{...}{for specific tests separeted by commas}
}
\description{
Monitor coverage of a package
}

