% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{spec_trunc_autocov}
\alias{spec_trunc_autocov}
\title{Spectrum-wise truncated autocovariance matrix estimator}
\usage{
spec_trunc_autocov(X, l, tau, ...)
}
\arguments{
\item{X}{A \code{numeric} p x n matrix of observations.}

\item{l}{A non-negative integer corresponding to the lag of the autocovariance.}

\item{tau}{A \code{numeric} scalar corresponding to the truncation parameter (larger than 0)}

\item{...}{Additional arguments.}
}
\value{
A p x p matrix corresponding to the spectrum-wise truncated lag-l autocovariance matrix estimator
}
\description{
Compute the spectrum-wise truncated autocovariance matrix estimator.
}
\examples{
set.seed(123)
X = matrix(rt(50*100, 3), nrow = 50, ncol = 100)
gamma_1 = spec_trunc_autocov(X, l = 1, 6)
dim(gamma_1)

}
\author{
Haotian Xu
}
