% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{VAR1_simu}
\alias{VAR1_simu}
\title{Simulating VAR1 model with different error distributions}
\usage{
VAR1_simu(n, mu, skip = 300, Sigma.mat, rho, err.dist = "t3", ...)
}
\arguments{
\item{n}{A strictly positive integer corresponding to the length of output series.}

\item{mu}{A \code{numeric} vector of mean of VAR1 model.}

\item{skip}{A strictly positive integer corresponding to the length of initial series not returned.}

\item{Sigma.mat}{A p x p matrix such that Sigma.mat^2 is the covariance matrix of the error term.}

\item{rho}{A scalar in (0,1) indicating the decay rate of the transition matrix for lag-k term as the k grows.}

\item{err.dist}{A \code{character} string, such as "normal", "pareto", "lognorm" or "t4", indicating the distribution of innovations.}

\item{...}{Additional arguments.}
}
\value{
A p x n matrix corresponding to the simulated data
}
\description{
Simulate a p dimensional VAR1 model with length n. The setting follows Ke et.al.(2019)
}
\examples{
n = 100
p = 50
Sigma = diag(1, p)
set.seed(123)
X = VAR1_simu(n = n, mu = rep(0, p), skip = 300, Sigma.mat = Sigma, rho = 0.7, err.dist = "pareto")
dim(X)

}
\author{
Haotian Xu
}
