% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{MonitorCoverage}
\alias{MonitorCoverage}
\title{Replace function body for coverage measurement}
\usage{
MonitorCoverage(func, package.name)
}
\arguments{
\item{func}{function object to be monitored}

\item{package.name}{optional name of the package for function replacement}
}
\description{
Decorate function body to be able to measure coverage. Function object is replaced in a proper namespace
}

