% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{MonitorCoverageHelper}
\alias{MonitorCoverageHelper}
\title{Generic pander method}
\usage{
MonitorCoverageHelper(stmt.list)
}
\arguments{
\item{stmt.list}{statement}

\item{...}{optional parameters passed to special methods and/or raw \code{pandoc.*} functions}
}
\value{
By default this function outputs (see: \code{cat}) the result. If you would want to catch the result instead, then call the function ending in \code{.return}.
}
\description{
Prints an R object in Pandoc's markdown.
}

