% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{StopMonitoringCoverage}
\alias{StopMonitoringCoverage}
\title{Stop monitoring coverage for a function}
\usage{
StopMonitoringCoverage(func.name)
}
\arguments{
\item{func.name}{}
}
\description{
Function stop clears the annotation added to monitor coverage
}

