% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcov.R
\name{StopMonitoringCoverage}
\alias{StopMonitoringCoverage}
\title{Stop monitoring coverage for a function}
\usage{
StopMonitoringCoverage(func, envir)
}
\arguments{
\item{func}{function or function name}

\item{envir}{environment where function with coverage annotation was placed. Primeraly used for debugging purposes}
}
\description{
Function stop clears the annotation added to monitor coverage
}

